create the update career details query rel #14
